requirements.txt
PyYAML~=6.0
click~=8.1.3
pytest~=7.3.1